dir: move to data_structure folder
- 관련 있는 모듈끼리 모아놓기 위해 자료구조 폴더 생성
